Fixed a comment bug

+++ Android/app/src/main/java/com/example/carousel/CommentAdapter.kt
@@ -27,12 +27,12 @@ class CommentAdapter (    private val commentList: ArrayList<Comment> ) : Recycl
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         holder.username.text = commentList[position].ownerID
-        holder.rating.rating = commentList[position].rating
+        holder.rating.rating = commentList[position].rating.toFloat()
         holder.body.text = commentList[position].body
 
     }
-    fun getRating(): Float{
-        var sum = 0f
+    fun getRating(): Double{
+        var sum = 0.0
         for(item in commentList){
             sum+= item.rating
         }
